+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Fixed bug in `db.Scripting()` queries (not checked operation results)
+* Added `sugar.ToYdbParam(sql.NamedArg)` helper for converting `sql.NamedArg` to `table.ParameterOption`
+* Changed type `table.ParameterOption` for getting name and value from `table.ParameterOption` instance
 * Added topic writer experimental api with internal logger
 
 ## v3.37.8